Add files via upload
A chessboard field is defined by a pair of numbers (a, b), each from 1 to 8, the first number defines the column number, the second the line number. Two cells are given. Determine if a chess rook can fall from the first square to the second in one move. Input data: 4 integers from 1 to 8 each, the first two specify the initial cell, the second two specify the final cell. The initial and final cells do not match. The program should output YES if the rook can be transferred to the second or NO otherwise from the first cell.